+++ packages/vite-plugin/CHANGELOG.md
@@ -1,5 +1,11 @@
 # @glissade/vite-plugin
 
+## 0.65.0-pre.0
+
+### Patch Changes
+
+- @glissade/core@0.65.0-pre.0
+
 ## 0.64.0
 
 ### Patch Changes